requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
flask-cors==3.0.4
numpy==1.18.1
